+++ packages/nimbus/src/components/toast/services/index.ts
@@ -1,8 +1,3 @@
 export { toast, ToastManager } from "./toast.manager";
-export {
-  getToaster,
-  getToasterEntries,
-  isToastersActive,
-  onToastersActivated,
-  resetToasters,
-} from "./toast.toasters";
+export { getToaster, getToasterEntries, resetToasters } from "./toast.toasters";
+export { isToastersActive, onToastersActivated } from "./toast.activation";
